+++ src/be/nikiroo/utils/test_code/SerialTest.java
@@ -75,6 +75,7 @@ class SerialTest extends TestLauncher {
 				encodeRecodeTest(this, data);
 			}
 		});
+
 		addTest(new TestCase() {
 			@SuppressWarnings("unused")
 			private TestCase me = setName("Anonymous inner class");